+++ packages/parcel-plugin-codegen/src/index.ts
@@ -1,5 +1,0 @@
-import type * as ParcelBundler from 'parcel-bundler';
-
-export = function (bundler: ParcelBundler) {
-  bundler.addAssetType('codegen', require.resolve('./CodeGenAsset'));
-};