+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.0 (tbd)
 
 - Adjusted for Piral 1.0.0
+- Use new codegen plugin
 
 ## 0.15.0 (November 17, 2022)